Update api.py (#18)
In `get_intel_notifications_feed` the GET parameter `page` must be `next`. Using `page` with the next token will return the same results (as if `page` parameter is ignored). `next` correctly gets the next set of tokens.

In `delete_intel_notifications` the list of ids must be sent in as a stringified JSON array. This is sent in as the `data` field in `requests.post()` instead of the `params` dict and the `api_key` is sent as a GET parameter in they URL as `key`. It may work another way, but we deduced this from analyzing the network calls while doing this on the website in Firefox's developer tools. Otherwise the response would always be `u'{"deleted": 0, "received": 0, "result": -1}'`.